Give the pattern-safety test its own class, with no fixture
WildcardPatternsUseTheNonBacktrackingEngine was added to PathAwarePatternTests,
whose constructor creates a temp directory, three nested folders and four files,
and whose Dispose deletes them. xUnit builds a fresh instance per test, so every
run did that filesystem work for an assertion that only calls CompilePatterns and
never looks at a file.

Beyond the waste, it gave a test that needs no disk a way to fail that has
nothing to do with what it checks: an IOException creating or deleting the
fixture would have been reported as a pattern-compilation failure.

It moves to PatternSafetyTests, which owns no fixture, matching the one class per
file convention the rest of the test project follows. PathAwarePatternTests is
byte-identical to its v3.14.0 content again, so the file that documents
-Include/-Exclude path semantics contains only tests about path semantics.

Verified: 756 tests still pass - the same count, so the test moved rather than
vanished - and it runs from its new class under a filter.
